refactor(web): naively extract module 'DesktopLayout' and 'Pages'
