+++ ui/web/src/index.css
@@ -10,11 +10,3 @@ body,
   margin: 0;
   padding: 0;
 }
-
-#root .flexlayout__tabset-selected {
-  background-image: none;
-}
-
-body[theme-mode='dark'] .flexlayout__tab {
-  background-color: #141414;
-}
